Refactor: Moved card identifiers to card div
This is so that it will be easier to flip over cards on an unsuccessful
match, because card divs with the class "flipped" and identifiers in
matchState.value1 and matchState.value2 can be targetted.  This results
in less DOM traversal, plus it makes sense semantically to have the
card identifiers on the actual cards, as opposed to the card containers.