feat: Add win condition
When the player makes 8 matches, the game is complete.  At this point
the player then receives a message alerting them that they have won, as
well as giving them the total of false moves that they have made.  The
value of false moves will be used eventually to determine their
finishing postiion on the podium.