refactor: Refactor reset() function
Move logic relevent to beginning a game to the beginPlay() function and
keep the logic in reset() that is only applicable once a game has been
played and the page needs to be "reset" before beginning the new game.

As part of reset() involves flipping over all cards face down,
beginPlay() is invoked using setTimeout which allows the animation of
the cards flipping to complete before the new cards are dealt.  The
interval set is 350ms which is half of the transition speed of a card
flip.